Added verbose flag and fixed attempts param

+++ src/cli.zig
@@ -21,18 +21,13 @@ pub fn printHelp() void {
     print("Required arguments:\n", .{});
     print("  --hands <number>       Number of hands to simulate\n\n", .{});
     print("Optional arguments:\n", .{});
-    print("  --bankroll <amount>    Starting bankroll amount (default: ${d:.2})\n", 
-          .{GameConstants.default_starting_bankroll});
-    print("  --minimum <amount>     Table minimum bet (default: ${d:.2})\n", 
-          .{GameConstants.default_table_minimum});
-    print("  --spots <number>       Maximum spots at table (default: {})\n", 
-          .{GameConstants.default_max_spots});
-    print("  --decks <2|6>          Number of decks (default: {})\n", 
-          .{GameConstants.default_num_decks});
-    print("  --attempts <attempts>  Number of simulation runs (default: {})\n", 
-          .{GameConstants.default_attempts});
+    print("  --bankroll <amount>    Starting bankroll amount (default: ${d:.2})\n", .{GameConstants.default_starting_bankroll});
+    print("  --minimum <amount>     Table minimum bet (default: ${d:.2})\n", .{GameConstants.default_table_minimum});
+    print("  --spots <number>       Maximum spots at table (default: {})\n", .{GameConstants.default_max_spots});
+    print("  --decks <2|6>          Number of decks (default: {})\n", .{GameConstants.default_num_decks});
+    print("  --attempts <attempts>  Number of simulation runs (default: {})\n", .{GameConstants.default_attempts});
     print("  --quit_threshold <amount> Stop when bankroll reaches this amount " ++
-          "(default: ${d:.2})\n", .{GameConstants.quit_threshold});
+        "(default: ${d:.2})\n", .{GameConstants.quit_threshold});
     print("  --strategy <strategy>  Betting strategy (default: increase)\n", .{});
     print("  --seed <string>        Seed for random number generation (optional)\n", .{});
     print("  --help                 Show this help message\n\n", .{});
@@ -88,7 +83,7 @@ pub fn parseArgs() !GameConfig {
             }
         } else if (std.mem.eql(u8, arg, "--attempts")) {
             if (args.next()) |attempts_str| {
-                const attempts = try std.fmt.parseInt(u8, attempts_str, 10);
+                const attempts = try std.fmt.parseInt(u32, attempts_str, 10);
                 if (attempts < 1) {
                     print("Error: Number of attempts must be at least 1\n", .{});
                     std.process.exit(1);
